introduce Item-derived Movie type and fix type inheritance
First, a metaclass is introduced to simplify handling of Blueprints,
most importantly using BPs defined in base type along with the ones
in derived type(s). This approach eliminates the need for caching
Blueprints and iterating over cls.__dict__ altogether, since all
BPs, no matter which class they are defined in, are cached at the
moment of constructing the new class. More advantages of this
system are to come.
Additionally:
* the new Movie type is introduced,
* all Item member functions are patched to use the new mechanism,
* UserData and RawAccess classes are patched as well,
* API is fixed for better clarity between class names vs objects.