Presenter configuration and serialization
Presenter now communicates with containers to get column widths and headings.
Presenter can be constructed from a configuration string which is a JSON dump
of a dict containing columns to be presented (as keys) and their configured
widths (as values). Widths equal to None signifies that a default value should
be used. Default values are now parts of Blueprints.
There's also a new defaults.py module to contain default configuration for
presenters of different types.