API session management & database updates
This commit introduces communication between FilmwebAPI and the GUI. When the
API object is instantiated, it receives a callback to the GUI-driven function
which opens the Login window, allowing the user to input their data, and then
resorting back to the API for the actual login.
Another new feature is the "first run". By default, API has no user name data
which means that anyone can log in. But when loading existing user data, only
this user's name can be used to update the database. So if the API is created
with some given username, it will notify the GUI (via the callback) that this
name cannot be changed by the user.
The API itself has been improved with a session management system. Its public
methods require the session and username to be present (e.g. to generate URLs
for their requests) and would fail if those are not present. They are guarded
by a enforceSession decorator, which ensures the session is present (if it is
not, it is requested using the GUI callback) before allowing the execution to
proceed, or forcing a null return. A short patch has been also implemented in
the Database code so that it can handle a login failure and, for example, not
lose all its data.
Hack-in and demo page methods have been removed.