overhauled filtering mechanism with an example: years
The new idea for filters is that they will bind together the GUI behavior along
with the backend operations. It's up to the Filter class not only to define how
does the filtering work but also how are the parameters retrieved from the user
(that is, how do the widget looks like and what it does). So every Filter class
defines a buildUI function (in which all the internal widgets are created, with
their backend relationships) which constructs a tk.Frame - essentially, Filters
are widgets that can be placed in the Presenter using TK interface (pack/grid).
Filter objects create callables which can be evaluated on Items, returning bool
values. To enable filtering with multiple criteria, every Presenter is equipped
with a FilterMachine, which holds control over the Filters and aggregates their
functions into a single expression which is then evaluated by the Presenter for
each Item on its list. Every filter, when it is altered by the user, calls back
to the Machine, which in turn calls back to the Presenter to update the display
with the new filter.
Adding a new Filter to the Presenter happens via the addFilter function, taking
a Filter *class* and the desired position within this Presenter's filter frame.
This method instantiates the Filter object (equipping it with the callback) and
registers it with the Machine, finally placing its widget in the window.
The first filter is introduced in this commit: a YearFilter, to selecting items
with the 'year' property from within a given range.