mechanism for populating Filter data with an example: genres
Each Filter now can also define a populateChoices method, which takes a list of
objects and may go through them to retrieve all possible values to populate the
widget data (example in GenreFilter code). When the Presenter makes a total up-
date, right after fetching the items from a Database, it makes a call to Filter
Machine, giving it the new objects. The Machine goes through all of its Filters
and executes populateChoices on them and resets them.
Minor improvements:
* fixed bug in filter ID assignment,
* adjusted the filter reset() interface,
* no more srcdataUpdate/totalUpdate - since "total" does nothing more than just
  calling srcdata, there doesn't seem to be a point in multiplying entities.