generalization of Filters based on tk.Listbox with an example: countries
An intermediate class handles construction of a scrollable Listbox, acquisition
of its data, management of its selection and resetting it. GenreFilter, as well
as the newly introduced CountryFilter, make use of this additional abstraction.
A considerable portion of old GUI-managed filter code is removed. The rest will
go away as soon as the remaining filters are reimplemented.