filtering overhaul done: DateOf filter, ResetAll button in Presenter
Clever bit: Presenter keeps track of the row and column position of the Filters
so that when all of them are placed in the filter frame, the ResetAll Button is
placed in the furthest place in south-east corner. Its placement is deferred to
the moment of first totalUpdate (only then it's certain that all of the Filters
have been added).
All remains of the old filtering code have been removed from the GUI along with
the old filtering.py module.