do not expose that much of the internal filter business to the presenter
The Filter constructor does not need to know about the FilterMachine callback as
the callback behavior is exactly the same for every filter and could be abstract
with respect to the implementations instead. So now, constructor only takes what
it really needs (i.e. just root), relying on the machine to give each filter the
callback to itself (instead of explicitly doing that in the Presenter code).
Also updated the technical document a little.