more intelligent Year and Rating filters
Now, when typing (or clicking) in some impossible filter parameter combination,
such as: from years 2005 to 2004, the Year and Rating filters will correct such
inputs by detecting which of the variables has just been modified and adjusting
the opposite one accordingly (e.g. if the user typed "2005" as the lower limit,
the upper limit will be automatically adjusted to "2005" as well, thus allowing
at least some items to appear).
The same was attempted for DateFilter, but the situation there is a little more
tricky due to the values having three components. Leaving as a "maybe to do" :P