auto-updating through github
An online update system is introduced in the updater.py module, via the Updater
class. Mainly, the system comprises three core components: background thread to
check update availability, a version comparison system (using semantic_version)
to compare the current program version against the available one, and finally a
downloading subsystem with checksum validation.
The Updater provides an internal thread for checking whether there is an update
available. To avoid problems on the boundary of TkInter GUI and threading, they
are kept somewhat separate. GUI constructs an Updater and calls its (relatively
lightweight) method checkUpdates, which does two things. First, spawns a thread
that does the actual online check (explained below). Then, it calls an internal
routine for checking whether the thread is done - this uses the after method to
call itself in a loop until some specific flag (self.checked_flag) is set. This
flag can only be set by the thread after it is done with its work.
Before elaborating on the actual update checking mechanism, a short explanation
of the version comparison system is needed. Every update will be accompanied by
a metadata file (called VERSION.json), containing the version number (following
the semantic versioning name scheme) and dictionary containing all of the files
belonging to the program at the given point (names as keys) and their checksums
(as values). With this in place, the client checking for an update can download
the metadata (this is fast) and compare the current version number with the one
available online - using the semantic_version module. Then, the thread sets the
check completion flag and the update availability flag (self.update_available).
One the next after() call, a dialog box is shown to the user so they can decide
whether to download the update.
Upon accepting, control flow is passed to the Updater, which downloads files on
a one-by-one basis, pulling only those files whose checksums have changed since
the last update. Checksums also allow verifying correctness of the download (to
minimize the risk of replacing working but old version with a corrupted one).
After updating, the program must be restarted (or have its modules reloaded and
all objects reconstructed - which is impractical), which happens automatically,
right after the update is done, without any additional permission from the user
(they have already agreed to update anyway). Updater knows GUI's quit method so
this is a simple matter of executing a callback. Main then calls to the system,
requesting the relaunch of the entire app.
The final contribution is the debug mode, designed to test updates pre-release.
Normally, the program pulls the metadata file from the master branch, then rest
of the data from a specifically tagged commit (a letter "v" followed by version
string). When launched in the debug mode, the program downloads metadata from a
separate, testing branch ("prerelease"). Rest of the update follows normally by
constructing a tag string and pulling files from there. The key difference lies
in the metadata download location - this allows pushing a test update to such a
location and testing it before shipping to the end users (who, by default, will
pull it from master branch).