simplified installation
The launch script has been streamlined to install dependencies. First, when the
program is launched for the first time, the script will install all packages so
that the user doesn't have to go through any extra steps. Second, after such an
update which added a dependency, this will allow installing it during a program
restart.
One tool added in this commit is the dependency_test.py whose task is simply to
attempt importing libraries and generate the list of packages that are missing.
When the launcher notices that such a list has been generated, it instructs pip
to digest the list.
The only drawback of this solution is that it is very Windows-centric. Users of
Linux will unfortunately have to do some of that stuff themselves.