cleaned up and streamlined release code
Tags are no longer needed for shipping, which is a good thing because they were
very error-prone (tag name had to exactly match the version string with a "v").
Now updates will happen entirely via branches. A stable snapshot of the repo is
going to be on the "stable" branch, while a debug preview on "prerelease". This
allows streamlining the update code: now the same thing will happen, regardless
of the mode (debug or not), the only difference being the repository root path.