intermediate layer for user data loading
A class dedicated for managing user data will allow easily maintaining backward
compabitility with older user data file standards (as with more features added,
the file layout might undergo changes). The system consists of a UserData class
(providing simple semantics for data access instead of clumsy retrieval of dict
keys), a DataManager class (the actual wrapper with load() and save() functions
that operate on UserData objects) and a Loaders class (stores all of the legacy
as well as current methods for loading a data file - "loaders").
When adding a new loader to the Loaders class, the information must be provided
about the data files versions it is able to process. Using a special decorator,
new loaders can be registered with the DataManager along with the earliest data
file version string they parse. During the actual loading of a data file, first
its version string is read, and basing on that, the correct loader is retrieved
from the registry and used to construct the UserData object that is returned.
TODOs: data saving, a way to handle constructing only the most recent format of
the UserData object even with legacy loaders.