reworked and simplified DateFilter with Comboboxes
No more need to take care of manual input or value wrap-around. Date components
can only be selected from a set of strictly controlled values. Switching to Int
variables eliminates the need to validate inputs. Attention shifts more towards
validating the actual date - e.g. what happens when the date is 2019-03-31 when
the user changes month to 2. Not only the resulting date needs adjustments, but
the set of allowable values to the day entry needs to be corrected as well. All
of such changes are now better organized: see _update, _getDates and _setDates.