test suite: basic online interface test & caching
Logging in and fetching raw HTML data from Filmweb is crucial,
therefore it needs to be tested first. We isolate this test in
a separate case class, and additionally cache its results (i.e.
fetched pages) so that more detailed tests (parsing etc.) don't
need the user to log in every time.