post-request session validity checking
Now that the session is persistent between runs, the assumption that a live
session will always be also valid no longer holds - session can get stale
and the backend will only let us know that after a request is made with it.
If a request requires a live and authenticated session but the current
instance is stale, backend will not return the expected data but instead
give a generic response containing a specifically named span. Therefore,
locating that span in the response serves as a good check for post-request
session validity check.